fixing filters and bugs
