+++ my-video-room/module/sitevideo/index.php
@@ -30,7 +30,7 @@ use MyVideoRoomPlugin\Module\SiteVideo\MVRSiteVideo;
 		->add_activation_hook(
 			fn () => Factory::get_instance( MVRSiteVideo::class )->activate_module()
 		)
-		->add_admin_page_hook( fn () => Factory::get_instance( MVRSiteVideo::class )->render_sitevideo_admin_page() )
+		->add_admin_page_hook( fn () => Factory::get_instance( MVRSiteVideoRoomHelpers::class )->render_sitevideo_admin_page() )
 		->add_deactivation_hook(
 			fn () => Factory::get_instance( MVRSiteVideo::class )->de_activate_module()
 		);
